coomons folder cleanup
